+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Remove `rxjs` and `lodash` from deps
+- Move `react` to peer deps, allow newer versions
+- Bump deps
+- Updated bundle
+
 ## 5.3.0 - 2023-06-22
 
 ### Added